requirements-lock.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
distro==1.9.0
fastapi==0.135.3
filetype==1.2.0
google-auth==2.49.1
google-genai==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.26
langchain-google-genai==4.2.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.26
mangum==0.21.0
mypy_extensions==1.1.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==1.0.0
structlog==25.5.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
